requirements.txt
streamlit
opencv-python
pytesseract
numpy
Pillow
ultralytics
